Reduce height of ListView to parent
If the ListView height was set to wrap_content, potentially
many views will be bound while loading. In practice it was
observed ~130 views were bound, when only several were displayed
at a time.

To reduce the number of views bound when loading, setting the
the height to the parent has the result of reducing the number
of views needed at a time, and half of the views are now
bound while loading.