finish budget view when moving to budget edit
The editing of the budget may result in its deletion.
To avoid attempting to load a budget that does not
exist, do not plan on viewing the budget after editing
it.

Further, the expected use case is to transition back
to the budget list anyway, so viewing the budget again
after editing was unnecessary.