Edit Budget/Transaction on long click
Feedback was received on the expected workflow of deleting or
editing a Budget or Transaction. It turns out the expectation
is to long click an item to reveal non-typical options. Clicking
on an object should reveal the most common option.

To this end, a long click on a Budget or Transaction will now bring
up a context menu, showing an edit option. Clicking on a Transaction
will still display it, but clicking on a Budget will no longer
do anything.