Reduce min SDK version to 12
The selection of SDK 17 was arbitrarily based on the version
available on my device at the time. As no APIs are being used
at that level, a lower SDK version can be targeted.

All the APIs currently used by the application were available
in SDK 12+. Moving the min SDK version to 12.