Launch expense view in new task from widget
Previously the widget would attempt to bring the
TransactionViewActivity to the front if it exists.
However, this may be in the state of viewing a transaction
or attempting to add a revenue. In these cases, the widget
did not bring up the add expense view which was expected.

As the widget is expected to only bring up a single-purpose
dialog, use the new task and multiple task flags. This will
result in launching the TransactionViewActivity in its own
task, and once complete will return to the home screen.