Allow user to configure jpeg quality for receipt images
Previously the receipt images were saved at whatever jpeg quality
was used on the images from the camera. This quality may be much
higher than is needed to store a readable receipt image.

To prevent receipts from taking up too much space, this commit
adds a setting for the jpeg quality to use for receipts. If set
to less than 100, the receipt image is re-encoded before being saved.
The default quality used is 40, which empirically seems to be more
than sufficient for representing a receipt with minimal image
artifacts.