Remove usage of parentActivityName attribute
In all cases the normal flow of activities onto the stack is the
expected order; there is no need to force a parent activity.
In the case where a budget is clicked, bringing up a transaction
view filtered by the budget, clicking 'back' resulted in going
back to the main view. This was not expected. Removing the
parentActivityName attribute resolves this. As a result, the
back button now needs to be explicitly handled.