Add ability to select file to import and send export data
To date the only way to import a file is to hope the exported file
is in the correct location. To remove this requirement, this change
checks if there are any activities available on the system which
can look in the file system. If there are, new options for
importing data will be available.

In addition, the import/export activity's layout was updated,
adding more explanation.

Finally, a Send... option was added to the export dialog,
allowing one to send the exported data to another application.

German translations will need to be added in a separate
commit for the new strings.