Add ability to search transactions
There is now a search icon in the action bar at the
top of the transaction view. When clicked it will
allow one to enter in a string to search. The string will
look through the description, account, value, and note
fields to see if the query string exists in any of those
fields. Each match is displayed.

In addition, the database methods for searching
for transactions was simplified so that it may handle
query filtering, expense/revenue filtering, and
budget filtering all in the same place.