Show all transactions when search is closed
If a user entered a search term, is shown the results,
but then closes the search box the expectation is that
the original list, i.e. everything which was shown
before the search, should be shown now. This change
adds this behavior by monitoring when the search
box opens and closes, and will refresh the list
upon a close.