Show data in TextViews when viewing
Earlier when an item was viewed all its data was shown
in an EditView which was disabled. This was not optimal,
as the resulting text was dimmed. To help the text be easier
to read, a TextView is now used to display the data.

In addition, the layout of the budget and transaction
views has been improved and standardized.