Use US locale when formatting floating numbers
When validating the value field in a transaction
Double.parseDouble() is used. This expects that
decimal marker be a '.' however in some locales
the ',' character is used. If the default local
is used to convert the value into a String a ','
is used, which fails the input validation.

To prevent this issue, all cases where a floating
point value is formatted the US locale is used.
This is to maintain consistency.