Remove context menu for transactions
This functionality is duplicated, as one can click
on a transaction and then click on the edit menu item.
Because of this, the context menu is unnecessary.

In addition, there is a bug where searching a list
of transactions, then editing one via the context menu,
returning, then attempting to edit again resulted in
the transaction list being empty. Removing the context
menu, which is unnecessary, removes the need to resolve
this bug.