Ignore FileUriExposedException for now
In SDK 24+ if a file:// Uri leaves the activity then an
exception is thrown. Instead, content:// Uri's should be used.
This affects the capturing of receipt images from the camera.
This should be resolved by using a FileProvider, however in
doing so there was difficulty in getting Robolectric to
successfully test the result. Until this is figured out,
downgrading this issue to a log message.