Allow keyboard to appear when adding/editing budget/transaction
Initially there was an issue where when one viewed
a budget or transaction (not to edit or add) the keyboard
would appear. To prevent the keyboard from blocking
the screen, the windowSoftInputMode was set to stayHidden,
which would prevent the keyboard from appearing until
something was selected. Removing this no longer results
in the keyboard appearing on the view screens. Perhaps
that behavior was a bug in older Android versions, or perhaps
the new layout does not cause the problem. Either way,
to improve the speed at which one can add a transaction,
the stayHidden setting is now being removed.