Export receipt images using FileProvider
Starting in Android 24 file Uris cannot be used
outside of an Activity. To allow images to be
captured with a camera, a content Uri needs to be used.
This change will use a FileProvider on Android 24+.
For earlier versions the file Uri will still be used.
This provides backwards compatability, as well as allows
Robolectric to continue to test the general code path.