Remove using ProGuard
On Android 7 it was found that the share button on
the receipt sharing activity was missing, because
ProGuard removed a needed class. This may not have
been caught when ProGuard was enabled, or it may
have been a changing requirement in Android 7.
Either way, ProGuard has caused a few issues thus far
with minimal benefit.