Add support for transaction add shortcuts
The widget allows one to enter the "Add Expense" view
quickly. However, shortcuts are equally valid and simpler.
This adds an option to the main screen's menu to add
two shortcuts, ones to add an expense and one to add
a revenue. When clicked, the corresponding transaction
add screen is loaded.