Do not resolve filename from Uri
It was observed by one user on another app that uses
the same code that the Uri to filename lookup
can fail with an IllegalStateException. The code which
converts a Uri to a file path may not be that great,
especially now that the app is expected to not receive
file Uris.

Instead of determining the best way to lookup the file
names, as it is not that important, the Uri is used directly.
This is both for printing the location after the import,
and determinimg the mimetype.